Reduced setups times by checking for presence of document before embed upsert

+++ auxiliaries.py
@@ -66,7 +66,7 @@ def batch_upsert(iterable, batch_size=100):
         chunk = tuple(itertools.islice(it, batch_size))
 
 
-def embed_upsert(filepath, verbose=False):
+def embed_upsert(filepath, verbose=True):
     """
     Takes a read-in file gets its embeddings and upserts the embeddings to pinecone vector database
     """
@@ -75,6 +75,7 @@ def embed_upsert(filepath, verbose=False):
     def clean_filename(pathname):
         filename = os.path.basename(pathname)
         filename = re.sub(r'[^a-zA-Z0-9]', '', filename)
+        filename.lower()
         return filename
 
     namespace = clean_filename(filepath)
@@ -92,6 +93,12 @@ def embed_upsert(filepath, verbose=False):
          pinecone.create_index('general', dimension=768)
 
     index = pinecone.Index(index_name)                          # Initialize index
+
+    # Check if Namespace already availabe, if so terminate.
+    stats = index.describe_index_stats()
+    if namespace in stats['namespaces'].keys():
+         return ["File already Present in Database. Ask Away!", namespace]
+
     index.delete(delete_all=True, namespace=namespace)          # Delete namespace if exists
 
     for document in tqdm(documents):
